+++ cpp07/ex01/main.cpp
@@ -6,7 +6,7 @@
 /*   By: upolat <upolat@student.hive.fi>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/06 17:07:17 by upolat            #+#    #+#             */
-/*   Updated: 2025/03/06 19:14:59 by upolat           ###   ########.fr       */
+/*   Updated: 2025/03/06 19:44:01 by upolat           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,7 +18,7 @@ int main(void) {
 	std::string arr2[5] = {"c", "plus", "plus", "is", "interesting"};
 	char		arr3[5] = {'H', 'i', 'v', 'e', '!'};
 
-	int len = 5;
+	size_t len = 5;
 	iter(arr1, len, print);
 	iter(arr2, len, print);
 	iter(arr3, len, print);
